refactor: move messaging code to umb feature package
