+++ service/src/main/java/org/jboss/sbomer/features/umb/consumer/MessageConsumer.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jboss.sbomer.messaging;
+package org.jboss.sbomer.features.umb.consumer;
 
 import javax.jms.Message;
