+++ service/src/main/java/org/jboss/sbomer/features/umb/consumer/model/PncBuildNotificationMessageBody.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jboss.sbomer.features.umb.consumer;
+package org.jboss.sbomer.features.umb.consumer.model;
 
 import org.jboss.pnc.api.enums.BuildStatus;
 import org.jboss.pnc.api.enums.BuildType;
