+++ service/src/test/java/org/jboss/sbomer/test/messaging/PncBuildTest.java
@@ -34,10 +34,10 @@ import javax.jms.TextMessage;
 import org.jboss.pnc.api.enums.BuildStatus;
 import org.jboss.pnc.api.enums.ProgressStatus;
 import org.jboss.sbomer.core.test.TestResources;
-import org.jboss.sbomer.features.umb.consumer.PncBuildNotificationMessageBody;
-import org.jboss.sbomer.messaging.JmsUtils;
-import org.jboss.sbomer.messaging.PNCMessageParser;
-import org.jboss.sbomer.messaging.UmbMessageConsumer;
+import org.jboss.sbomer.features.umb.JmsUtils;
+import org.jboss.sbomer.features.umb.consumer.PNCMessageParser;
+import org.jboss.sbomer.features.umb.consumer.UmbMessageConsumer;
+import org.jboss.sbomer.features.umb.consumer.model.PncBuildNotificationMessageBody;
 import org.junit.jupiter.api.MethodOrderer.OrderAnnotation;
 import org.junit.jupiter.api.Order;
 import org.junit.jupiter.api.Test;
